Implement standard stream.Writeable interface for write/end
Implementations of `stream.Writeable.write()` and
`stream.Writeable.end()` are supposed to take optional encoding and
callbacks. (In the case of `end()`, the `chunk` itself is also
optional.)

Adding a callback also increases reliability in the case that a message
is sent as the stream is closed because it will wait to close the
underlying gelfling client until the message is actually written.